autodoc shouldn't assume there's a converge playbook
When overriding the default molecule playbooks, the converge playbook
might be missing. In this case, we should refer to the molecule
configuration.

Also, when running a molecule test, normally the
MOLECULE_SCENARIO_DIRECTORY environment variable is defined. We need
to manually define it as it might be included in the playbook paths.

Change-Id: If641ec3547df9a452c85eb7e9256393d826c6541