Cast tripleo_cephadm_container_tag to string
The cephadm container image tag is usually alphanumeric (e.g. it can
be found in the form of <major>-<minor> or something like that).
For this reason we need jinja to produce the right string that should
be passed to the --image option and avoid an unexpected templating
error where the tripleo_cephadm_container_tag must be str and not int.

Change-Id: Ibaf94f72d3532ebac580cca96f71ad8d25ff2eff
(cherry picked from commit 8bab831a8a1c543adc1cc99d2a28135590ac0551)