tripleo_firewall: Allow injecting frontend rules
This change introduces new variables to define the firewall rules for
haproxy frontend.

When an API service is enabled, we should add proper firewall rule not
only in the node where the API service is running, but also in
the loadbalancer node where haproxy is running, otherwise the frontend
port in haproxy is not accessible.

The new variables are used once [1] is merged.

[1] https://review.opendev.org/c/openstack/tripleo-heat-templates/+/831549

Related-Bug: #1961799
Change-Id: I9d79df8a8d0eaf77166b296178b9b0622263998d
(cherry picked from commit f2760f5de319c07b233ea6821ef366601e557558)