Ensure english locale when running the command for ovs checks.
We parse the output of commands to get information around the
OpenVswitch versions. At least one of those commands supports
localization[1]. In order to ensure that the parsing won't fail on
unexpected character, we have to make sure all commands run with an
known locale.

We add a wrapper around the module.run_ansible ansible ansible command
that ensures that all commands run with C.UTF-8 locale.

We need to support string and list for that wrapper for command
because run_ansible[2] has a different interface depending on the
command being a string or a list. We use this behaviour to get access
to the shell.

[1] The dnf command.
[2] https://docs.ansible.com/ansible/latest/reference_appendices/module_utils.html

Closes-Bug: #1964155

Change-Id: I60f6efebecf24f3c3337cc02e8c61e0835de6ed0
(cherry picked from commit 0cca94e64f00e6cee382f05b54ea14c632dd4b29)