bnr - Fix bad UEFI working
In the past, some code was introduced on the B&R automation to take some actions in case the UEFI nodes had a certain boot configuration.
It was verified that the boot configuration was not related at all with how OpenStack manages UEFI, but it was an error on the physical installation of the OS.
That code does nothing so it is cleaned up with this change.

This change solves BZ/2020656

Change-Id: Ib9af92fbe9e4a7ed55291ebf7ed019129609daa6
(cherry picked from commit 1ad4396423fe59b4ce3bae7ca5e1e4217f02555c)