Move creating empty ruleset for deployed server
We create empty firewall ruleset for depoyed server in
tripleo_bootstrap and then create/save the firewall rules
in host_prep_tasks. If there is any failure in deployment
before tripleo-firewall tasks are run, then we end up with empty
file and rebooting any compute node would result in no
rules. If the intent is to cleanup the existing rules before
we save the new rules, let's do that before we create the
rules.

Also it looks like ipaddr filter has moved to ansible.utils
collection, that results in molecule job failure.

Change-Id: I31a5500e127099a6f7c23791543bc262e1715cea