Use node UUID for bridge_mapping agent check
When we check for bridge mapping. If the user
has provided the Name of the node, rather than
the UUID, then we will fail to lookup the agent.

This change ensures that we lookup the agent using
the nodes UUID rather than the name, regardless of
what the user has input in the provide command.

Closes-bug: 1966155
Resolves: rhbz#2067261
Change-Id: Ia6c7c93525d99df457482cda11e996656555f5e1
(cherry picked from commit a3c700e12da5efaaac31e4ccd8c2167f6ad0617b)