Improve connectivity check
Currently, the ICMP check can fail if the network
configuration takes too long. This can fail the
overall deployment, but the network configuration
can still be successful. This change broadens the
scope of the connectivity check by increasing the
number of ICMP requests sent each time. To make
it more flexibile and allowing for situations where
interfaces take a little longer to become active.

Change-Id: I89a3b9a562fb060293deb08dbc7e97aacfcc5b73