Continue providing remaining baremetal nodes for additional failures
Currently it continues processing nodes even it fails to start
providing some nodes unless failfast is false, while it immediately
fails and suspend remaining operations if it fails at an earlier step
like unlocking a baremetal node.

This change allows the module to continue processing the remaining
baremetal nodes, because the failure might be specific to only some
of the target nodes.

Change-Id: I628b5a653e9deacbfda120300198513efde0062c