Pass CephClusterName in deployed Ceph template
If the user overrides the tripleo_cephadm_cluster variable,
then we should pass the same value through to the deployed
Ceph template to ensure that it is consistent during overcloud
deployment.

Also, make ApplyCephConfigOverridesOnUpdate the last parameter
in the template so that it is easier for users to see and follow
the commented recommendation.

Related-Bug: #1966559
Change-Id: I4daab327c0bdac2bcc569cc1d8d629c648ddf292