Remove missing stat warning
When we run a task with any delegation (often) the profiler is not
able to store or access stats, which raises a warning within our profile
callback. This change modifies the warning output to use debug instead
which cleans up our UX by removing a lot of warning output from a given
playbook execution.

Change-Id: Ib31b8de51b341502f39d135ee0f034bc052fb18a
Signed-off-by: Kevin Carter <kecarter@redhat.com>