[RFE] Add a new Snapshot And Revert feature to Backup and Restore
This RFE improves the B&R capabilities to better support the
upgrades context. Before an upgrade, it gives a lot of security to the
user to have a backup that can be recovered whenever something goes
wrong or the upgrade window closes. The main requirement for upgrades is
to be quick, both in the backup and in the restore phases, but most of
the B&R provided features (such as storing long-lived backups on a
different machine) are taking a lot of time and they are not useful for
upgrades use-case.

In this context, we developed the Snapshot and Revert feature that
configures an LVM-enabled system to use LVM snapshots.

Change-Id: Ie355206045f430d1353b1557f43bd99c85edd86e