Check whether an HA resource already exists explicitly
With ephemeral heat we lost the meaning of the 'stack_action' hiera key
which we previously used to distinguish between fresh deployment and
pre-existing deployment (aka redeploy).
Since this hiera key is not available anymore, we simply check if the
resource existed before running puppet. If it did not, we're part of
an initial deployment, if it did exist then this is a redeploy and
in such case, if the config changed we will restart the service on
the bootstrap node.

We also cleanup the molecule steps as there are some redundancies there.

Testing protocol is in the parent change.

Note: manually squashed I5199cf5fd20590a2aa1af5c52ef805067f995366
to fix restart of all HA resources without stack_action.

Related-Bug: #1942309
Change-Id: I08cff959d7d104ad0e8d833ed6ce06df8ff38b16
(cherry picked from commit b8d9370997ee6f910d401399f72f97272d639c73)