Undercloud mysql database must be idempotent
Currently, the database backup excludes certain tables because it is
intended do a restorable dump of a Galera cluster, but this does not
apply onm the undercloud database, which does not belong to any cluster.

This patch changes the procedure for taking an undercloud database
backup to get a full one.

Change-Id: I5e2876b3f7ffe2b63584802505a10dd1966be4a0
(cherry picked from commit 7884d62ae8643bfa236a424cc9b90a819c2fd167)