ReaR needs specific configuration in RHEL9 with LVM
As RHEL9 has changed both the ReaR version (up to 2.6) and the provided
images are using lvm by default, both ReaR and the OS need little tuning
to be able to work. This patch leaves ReaR and the OS with the
appropiate configuration to work on RHEL9.

Change-Id: I49adc830bf9f4a32832a565214e4a586c514a988
(cherry picked from commit f04924812a8950b1b5d9cf06761847248cdee9e2)