Define role specific defaults for growvols_args
The Controller role will sometimes include swift object storage, and
the ObjectStorage role always will. This means unless the /srv volume
is grown it will be too small for swift to store any useful amount of
data.

The node provision tool has no knowledge of the specific services
deployed in a role, so it can only set growvols_args default based on
the role name. This change makes the default growvols_args role-aware
so that Controller and ObjectStorage roles grow /srv by enough to be
useful.

The documentation will encourage the deployer to consider actual /var
and /srv storage requirements (especially for the Controller role) and
setting custom growvols_args to reflect that.

It is impossible to choose defaults which meet all scenarios because
there are 3 unknowns (workload requirements for /var and /srv, and the
size of the node primary disk) so some discusson of Controller
requirements especially is encouraged.

If /srv or /var are deployed with insufficient space the deployer can
always add a second disk and expand the LVM volume into that, so there
is a mitigation option.

Change-Id: Ifaad8b5d6d243be1d2fa04ee028435e19b29098f
Closes: rhbz#2066367
(cherry picked from commit 7d840f551062f63f1241a1da992137aaeb5f842a)