Ensure the dependent service is started in advance
Currently when dependent service is defined using the depends_on key,
tripleo-ansible adds 'Wants' so that the dependent service is also
started. However this attribute does not guarantee that the dependent
service is started BEFORE the target service is started.

Because current usage of depends_on expect the dependent service is
started before the target service, this change adds 'After' so that
the ordering is properly enforced.

Closes-Bug: #1970354
Change-Id: Ic4ddbaba8aa8ca9501a00f5e4741fef73ba35b9d