Normalize the server_addr dashboard backend
The server_addr dashboard backend should be the same argument used as
instance of the current ceph-mgr container running on a controller node.
Due to a recent change in pacific the container name and the ceph-mgr
instance are not consistent anymore, and the purpose of this patch is to
rework the way this info is retrieved from ceph.

Closes-Bug: #1972729
Change-Id: Ifc9a0a9ac2f13c891ccde826aef2ab7cbdb5d690
(cherry picked from commit 0951120a9fecc169269515577b5ad6145db63157)