[TripleO-Package] Remove Dockerfiles
https://review.opendev.org/c/openstack/tripleo-ansible/+/839319
recommends use of default config_podman.yml config for podman driver
and removes the dockerfiles.

These are the leftover dockerfiles of tripleo-packages role.
This patch removes the same.

By using the same hostname, it fixes issue with the
tripleo_packages mol job
```
msg: 'Failed to connect to the host via ssh: ssh: Could not resolve hostname centos: Name or service not known'
```

Signed-off-by: Chandan Kumar (raukadah) <chkumar@redhat.com>
Change-Id: I2a7c76a82a4575e45cb0fc5031d8fe72858a7e6b