Add missing "inet nat" table
With the masquerade rules, we need to get the "nat" table in the "inet"
family so that we keep the overall consistency with all the other rules
and the dedicated chains.

This patch also correct an issue with the "daddr" and "saddr" parameter,
since we need to set what ipversion is used (ip, or ip6).

This allows to get the rules generated/used on the Undercloud when we
toggle the "masquerade" subnet parameter.

Change-Id: Ie998c4b892f59fe1909b9d0a53ee6b4059419260