Omit pg_num unless provided in tripleo_run_cephadm's build_pools
It's no longer required to pass a PG number when creating pools so
don't assume the pg_num will always be set. If it's not set then
handle it gracefully by omitting it so the pool is created without
a pg_num. The tripleo_cephadm role's pools.yaml tasks file already
does this.

Change-Id: Ief1a40161a1c57be8fd038b473a6f21feb9aa8fc