Use lowercase hostname in inventory network config
The ansible inventory uses the neutron port dns_name as
hostname. The `dns_name` in neutron is always lowercase.
If a hostname with one or more capital letters are used in
baremetal deployment YAML the data lookup from `inventory-
network-config.yaml` does not succeed.

This change fixes this issue by converting the hostname to
lowercase when creating the content for `inventory-network-
config.yaml`

Closes-Bug: #1978356
Change-Id: I1ee53976ecc63d206d9604c896b1bb59807ad9ff