Add ssl spec parameter to rgw
When tls-e is enabled, we need to make sure that the ssl keywork is
passed to the rgw spec. By passing this parameter (and setting it
to True), the tls mode for rgw is enabled, and the rgw_frontend_ssl
parameter is assimilated.

Change-Id: I92a35532e9a26a0f8d00660801ee8e8d5a329b8b
(cherry picked from commit 1e353efab254c8c5b0fad412d6fd352f8a4abf34)