[B&R] More robust check on Galera desync the MySQL node
This task will fail if anything is printed on the stderr, even
if it's a warning not related to the actual desync task. With
this patch you can set tripleo_container_cli_flags = '--log-level=error'
to supress this warning and continue with the backup.

Change-Id: I63063e6dd655b974ca2231eb55e03eb30c1461dc
(cherry picked from commit 32769ae9f9b18f64cce37c73bfa829f35e54fef0)