The stderr of the command can have WARNs
There is a task that it is re-tried until there are no more errors, due
to the fact that sometimes it does not work well the first or second times
because the DB has not been completely initialized. To achieve the
retry, we look for an empty stderr, but under certain circumstances
there can be WARN messages. We do not want to retry the execution after
a WARN message, as the action has been already succesfully executed.

Change-Id: Ia6053bf009fc13be1ea319e72c81b084bf3f0e9d
(cherry picked from commit a66c2b4943f8ee43352a449966d6dc4f8a9a74f4)