[update] Tell subscription-manager it's not running in a container.
Due to issue in podman[1], some file on the hypervisor might fool
subscription-manager into thinking that it's running inside a
container.

We hardcode the fact that it's not so that we are not impacted.

[1] https://bugzilla.redhat.com/show_bug.cgi?id=2097694

Change-Id: Idd526622c2c7ecf84ecde23311c5fbae9be82a0b
(cherry picked from commit 49658db7be3b10a2ae20a2b8cad6205fd7d4b4f9)