Only run Ceph network_config_set during initial deployment
There is no reason to include the network_config_set tasks
file in the cephadm_config_set tasks which can be run both
before and during overcloud deployment. It is safer to only
run it during the initial Ceph deployment. The danger is that
the network_config_set tasks file from the tripleo_cephadm
role can misconfigure Ceph during overcloud deployment if
the Ceph deployment relied on an initial ceph.conf to
configure the Ceph networks.

Change-Id: I74e1197362517c3ad4bbdfc54adbf3ac934acf4b
Closes-Bug: #1980992