Modify run_once behavior of tripleo_free strategy
It's very difficult to ensure that tasks are written without
run_once, expecting that they would be run with tripleo_free.
As tripleo_free does not ensure that the tasks are successfully
executed with run_once, let's go back to original ansible free
strategy behavior of running it for all hosts.

Change-Id: I1c67885f553aee519466249ac056f75071014c24
Resolves: rhbz#2061393