Support auto-revert of all the snapshots
There should be an action to revert all the snapshots at the same time
to provide a quick revert action, specially if something has gone
wrong during an upgrade.

At the same time, we divide the whole process into different tasks that
can be chosen to be executed via tags.

Change-Id: I70bfb00a704a1dba15be766e3a33d97f83912d6d