Filter out nftables keys when using iptables
Some of the nftables parameters are really only for nftables. Adding a
filter to ensure we can pass such content, even if using iptables, will
allow us to not break everything.

A good example of the usecase is for the logging - the iptables-nftables
doesn't convert any "log" jump to nftables, mostly because nftables
syntax is really different, allowing to provide more context.
In order to keep the tripleo_firewall role compatible with both,
introducing a "nft_" prefix to the rule parameters will allow to ensure
we don't break the iptables support.

Change-Id: I3f3e0957da4dbc6ccd4bc4d870e0b21c91973aeb