Preserve action: accept in tripleo_iptables
When using Puppet for firewall rules[1], we allowed
action: accept. This isn't something supported by
the ansible iptables module. However, neither is
action: drop, but we allow it within the
tripleo_iptables module. This change adds handling
for the "accept" keyword in the action paramater.

We will default to append in this case, rather than
insert.

[1] https://github.com/openstack/puppet-tripleo/blob/stable/train/manifests/firewall/rule.pp#L43-L45

Closes-Bug: 1982411
Change-Id: Ida74957a293e1b6097a7ca33e161f0fe0a7fd535