Configure podman netavark network - enable IPv6
In RHEL9 the default container networking used is "netavark".
With this change the default "podman" network is configured to
enable IPv6 networking for containers.

The default IPv4 IP address used is updated the match the ones used
in tripleo with "cni" networking.

Also, when "netavark" is the backend skip the tasks that do cni related
configuration.

Closes-Bug: #1983121
Depends-On: https://review.opendev.org/c/openstack/python-tripleoclient/+/851597
Change-Id: I90efb1639d7ffd899ab73446620354403c1ca536