tripleo_auditd: Clear unused default of tripleo_auditd_rules
This is follow-up of 3ed814e21fba7ca2328cd125c83d9941008a4583 and
replaces default for tripleo_auditd_rules by {}. This variable is set
to {} by default in tht and the current default value is not actually
used.

The current implementation uses a separate rule file (tripleo.rules)
and these defaults are already defined in the default file
(audit.rules).

Change-Id: I4106c1418e34518daaae895f2053019aec47e028