Unbound: remove incorrect config of forwarding
Configuring forwarders is only valid if
tripleo_unbound_forward_resolvers is defined and should also never be
set to bind servers that are meant to be authoritative.

The intention is that the forwarder configuration only be generated if
tripleo_unbound_forward_resolvers is configured with actual values.
This patch includes a fix to the conditional to skip creating the file
and also removes it in case it the deployment parameters no longer
include tripleo_unbound_forward_resolvers on redeploy.

Change-Id: I116253b60df1dc41237e06db9f6938c2bd3edd2e
(cherry picked from commit e858bbdd8d237da72bf8bd1c3794dc7c11df1827)