Add missing keyrings when ganesha is deployed standalone
When Ceph is external but ganesha is deployed by TripleO the Ceph cli
should be created accordingly and the keyring should be rendered in the
expected location. This patch adds the missing steps to copy and render
the manila keyring in the right locations and fixes the cli generation
to properly handle this use case.

Related-Bug: #1986988
Change-Id: I2651c2850debd8110da93df2adc5fd8768a00db0
(cherry picked from commit 96afe1302ca0959f80558bf0c135a9978bc65b0b)