Simplify dport writing, transform port list syntaxe
Make the rule generation a bit easier to follow, and avoid some code
duplication.

Also correct how the "port range" is written - nftables wants an hyphen
while iptables used a colon as a separator.

This change also enables passing multiple kind of data, such as:
dport:
  - 22
  - 8000:8100
  - 23

This will generate the correct {22, 8000-8100, 23} we need.

Change-Id: Iff57f0a07227dd188e74c77767de2f26dea14725