Add observability_data_only condition to node-exporter component
There are use cases where we'd like to take advantage of the unified
monitoring model in openstack, and the ceph cluster should provide a
set of deployed node exporters, but leave the data collection job to
a different component.
This patch introduces a new condition to allow a partial deployment
of the ceph monitoring stack: only node-exporter(s) are deployed
everywhere, but prometheus/grafana/alertmanager are skipped.

Change-Id: Ia1723394dae5ed03e50f2a236eb808f234e63331