Allow role re-tagging of ports
We currently tag ports with 'tripleo_role=<role>' and use
it for inventory generation. This does not help when we want
to move a set of nodes to a different similar role. This change
re-tags the ports with the new role, if there is a change.

Change-Id: I02e9aa5180d8262a6cb69933094526d5dbde8318
(cherry picked from commit 9ccc1b16d3fe908f70da5227d4804963092a21d8)