Do not execute anything if a host is unreachable
If a host is unreachable, the default behaviour for
ansible is to ignore that and execute all the tasks
in the hosts that are reachable. This is unwanted
behaviour for B&R, as we need the whole controller
cluster to be healthy and reachable before we take
a backup from it. This patch introduces a mechanism
to abort execution if there is one or more
unreachable hosts.

Change-Id: I3820875e72a8d805d71c7981468398bbcec0bf04
(cherry picked from commit a7f54163cf1bd5191431f7c4ca9a180b1a59ac58)