Improve and correct nftables role
Add a "cleanup" task file in order to get a clean environment; this is
especially nice for molecule, since it will allow to clean everything
existing during the "prepare" step, and at the end of the "converge".

Removes a typo leading to an unwanted newline in the generated ruleset;
this issue was introduced with [1].

Ensure we're compatible with puppet-firewall module. In there, the
"jump" was actually called "action"; the current template didn't take
that into account, leading to a potential issue if people were still
using the "old" (and wrong) notation.

More over, some more molecule tests have been added in order to ensure
ruleset is correctly generated - this should prevent future regressions.
Note that, for now, the tests don't cover *all* the usages, only
specific things that were raised during CI runs and testing.

[1] https://review.opendev.org/c/openstack/tripleo-ansible/+/855314

Change-Id: I91e68b7d237589fedf020dda71e1d987c43f11d3