num_dpdk_interface_rx_queues support net conf schema
The network_config schema in the baremetal provision defintion
did not allow setting the 'num_dpdk_interface_rx_queues'.

This change adds the varaible to the schema, and adds the option
in module: tripleo_generate_inventory_network_config. By default
the value is 1, which is the same default used in THT for parameter
NumDpdkInterfaceRxQueues.

Closes-Bug: #1989593
Change-Id: Ic28add116c369979b4038ce403451dc95dc2f53d