Fix typo when nova pool is built and wrong flag during cli creation
When the upgrade from nautilus to pacific is performed, the regular
cephadm playbook fails during pools creation. This is caused due to a
typo in the build_pools task that processes the vms pool. In addition,
during the same run, if no spec is present, post.yaml fails getting the
right cli due to a wrong 'mount_spec' flag value. This change fixes both
failures removing the extra space that caused the pool creation failure
and the mount_spec flag that caused an issue in the get_cli task.

Closes-Bug: #1990195

Signed-off-by: Francesco Pantano <fpantano@redhat.com>
Change-Id: I253c8feb72404e2620821472447b08905d5cd16b