Container build: define ARGs after FROM line
According to the Dockerfile reference[1], ARG instructions before the
FROM line are a special case: variables defined in this way are only
available in the FROM line and not during the rest of the build.

In order for defined ARGs to be available during image build time we
must run these after the FROM line.

[1] https://docs.docker.com/engine/reference/builder/#understand-how-arg-and-from-interact

Depends-On: https://review.opendev.org/c/openstack/tripleo-common/+/858864
Signed-off-by: Miguel Garcia <mgarciac@redhat.com>
Change-Id: I8086b7820ff2d124baacbe8e947aba970996b907