bnr - Update on rear scripts to support the ceph bnr on LVM
Due to the inclusion of lvm and the creation of different logical volumes, the backup and restore scripts for the /var/lib/ceph folder need to be adapted. It is required to change the directory path.
Closes-Bug: rhbz#2118397 - https://bugzilla.redhat.com/2118397
Change-Id: I5ad1f2c968be5087b71d7db2626ea96fe4959b25
(cherry picked from commit 3804820bf7cb950c2c5e85e66232e0cd5deb8e1a)