Update tripleo_nova_libvirt to virsh-secret define for ceph
Update standalone-ansible tripleo_nova_libvirt role to
handle the creation of the libvirt secret that is required
for nova to connect to volumes on cinder RBD backends.

Uses nova_libvirt_init_secret.sh from tripleo-heat-templates
(THT). This script is modified so that it can create the
initial secret so that Puppet does not have to.

A subsequent THT patch which depends-on this one can call
this Ansible role in its place.

Change-Id: Ieef097d343f4dd9aaec27b04d22e19d724fb5c6d