Add a new option for the vhost
We can now specify different Location, with a subset of options. This
will be needed for some of the service vhost, such as Keystone[1].

The layout is pretty simple:
tripleo_httpd_vhost_locations:
  '/my/location':
    - optionName: value
    - optionName2: value2
    - ...

Of course, a new molecule scenario has been added in order to ensure
this new feature is working properly.

[1] https://docs.openstack.org/keystone/yoga/admin/federation/configure_federation.html#configuring-an-httpd-auth-module

Change-Id: I553c1890864f63a32f23d9914df5490ed2577ecb