Clean tripleo_iptables related resources
Now that we've switched to nftables, it's time to remove
tripleo_iptables related resources. We don't want to support both
interfaces, since nftables is the future of the NetFilter interface (for
now).

We remove most of molecule tests as well since we're already testing
nftables directly in tripleo_nftables. tripleo_firewall is now a wrapper
ensuring proper service is enabled, and rules are properly listed.

We keep tripleo_firewall since it allows a smooth migration to whatever
engine the future will provide.

Change-Id: If9bdb915b28e33afd9ec2a487ccad5de3498188b