Split tripleo_nftables tasks
It will make integration into standalone playbooks easier. We also keep
the "main" for backward compatibility reasons.

We also take the opportunity to empty nftables if we're moving from
iptables to nftables; this ensure nothing will interfer.

This action is only done if we switch the firewall engine, meaning,
mostly, during a fresh deploy, or an upgrade. If it happens during an
upgrade, the services using the firewall will already be shut down,
therefore there won't be any weird behavior.

Note: this patch also changes the file extension to match molecule need
to get only .yml, not .yaml. Since the "cleanup.yml" is called in
different places in molecule, this patch must update multiple places at
once.

Change-Id: I761bcf06b01e7890182ec610ca4345c1ecb5707b