Improved way of detecting Galera cluster is up&running
There is a script in every node that checks correctly if the
Galera node has been correctly re-introduced in the cluster
as the previous check on the local port open had some
disadvantages, and does not wait for IST to complete.

This is an unclean cherry-pick because master have a couple of
conditionals that do not exist on wallaby.

Change-Id: Id49274f5bade4af5b88b6354cfa092e4b041eb6e
(cherry picked from commit dc4e574ba10807ea23155017c0bd9e734023fae5)