Only do snapshot on inactive nodes
The nodes must be taken out from Pacemaker cluster before taking a snapshot, one node at a time to not cause any downtimes. This will ensure that the database and the Ceph snapshots are consistent and we do not take the snapshot just during a write.

This behaviour can be deactivated using a parameter.

Also added the molecule data to use puppet lookup command

It is not a clean cherry pick. It was needed to modify the prepare.yml file to pass the molecule job because the role env_data is not present in wallaby

Change-Id: I3ea96ff2219733b3d8dd38680f2fa9eb50bb546b
(cherry picked from commit 626d4e1c99200faf03f27dd861daa855187d150c)