Don't use meta:end_play after growvols check
According to the documentation, meta: end_play will apply to all
hosts, which is fine if every host is deploying the same image.
However if there is a mixture of overcloud-full and
overcloud-hardened-uefi-full the playbook won't run correctly on some
hosts.

This change moves to a conditional block to run as expected on each
host.

[1] https://docs.ansible.com/ansible/latest/collections/ansible/builtin/meta_module.html

Change-Id: Iccda95cde0e309af00ce6b918ab1acab68e548dd
(cherry picked from commit 1678c3395ba8142e42ee1629bd1d258ffbdf4da1)