Fix compatibility with Ansible 2.14
This fixes the following three compatibility problems with the latest
Ansible 2.14 which was recently released.

1)
ITERATING_* attributes and FAILED_* attributes of PlayIterator has been
deprecated[1] and was removed in Ansible 2.14.

2)
The _valid_attrs attribute was replaced by the new FieldAttributes in
Ansible 2.14[2].

3)
The warn parameter of command/shell module was deprecated was removed
in 2.14[3].

4)
Drop usage of removed compat attributes from default callback[4].

[1] https://github.com/ansible/ansible/pull/73908
[2] https://github.com/ansible/ansible/pull/74511
[3] https://github.com/ansible/ansible/pull/77411
[4] https://github.com/ansible/ansible/pull/77677

Closes-Bug: #1996482
Change-Id: I74311e4c11b9dbb4858f119e109acfdaa8e81ff0