Mass corrections of molecule scenarios
With [1] in play, we're able to get a proper overview of the molecule
state. Apparently, some jobs were broken, and could merge because they
weren't run back then (there was some issue in the way zuul launched
molecule, running only the default scenario).

We also ensure we're cleaning dnf cache in case we're switching to
the mirror.

This patch ensures everything is square and straight so that [1] can
merge properly.

[1] https://review.opendev.org/c/openstack/tripleo-ansible/+/864994/

Change-Id: I8dfcce68a00ed4fd4a2f0ecbf094dc3689423dfd