Introduce tripleo_ceph_client_files standalone Ansible role
The tripleo_ceph_client_files Ansible role was designed to be used
standalone without tripleo-heat-templates or config-download. The
role copies files from a source directory on the host where Ansible
is run to a destination directory on the overcloud nodes. The file
permissions on the destination are set appropriately if file is a
cephx key or a configuration file.

The source directory needs to be prepopulated with Ceph configuration
and cephx key files.

Closes-Bug: #1994148
Change-Id: I06bfb7347715b2ae5a1b37d633ec66f1b069225a