[TRAIN-ONLY] Rotate ceph_ansible_command.log
Before calling ceph-ansible, check if the log file
ceph_ansible_command.log exists (from a previous run).
If it does, then rename it to append a timestamp. When
ceph-ansible is run a new log will be created. As per
the depends-on patch, the timestamped files will be
cleaned up after LogrotatePurgeAfterDays days.

This patch goes directly into Train because the modified
file has been removed in Wallaby.

Change-Id: I61f80b11c9394f34e09297ca1aa93d7000bb8047
Resolves: rhbz#2144485
Depends-On: Ifbd990f15b9d85260ab20a2004fe79bafda929d9