Fix missing links for old names
We have renamed the task files for tripleo_podman_role[1] but we missed
adding link for the following files though these are still used by
the collectd service template.
 - tripleo_service.yml
 - tripleo_rsyslog_cleanup.yml

This change adds the links to unblock CI now. All symbolic links will
be removed later once references in tht are updated to use the new
names.

[1] 0692c49525c3a5a4680ff0b9e15bbfe2822639a3

Closes-Bug: #1997664
Change-Id: I0646f49a6572b800b87a01f904833c689dc8759b