Handle child groups when building ceph spec from inventory
If the inventory contains the following:

  CephStorage:
    children:
      overcloud_CephStorage: {}
  overcloud_CephStorage:
    hosts:
      ceph-0:

and I want to get a list of hosts from the CephStorage group,
it is wrong to return that CephStorage has no hosts. Instead
since all hosts in the overcloud_CephStorage group are children
of the CephStorage group, I should return the hosts of the
child group. I.e. the hosts of CephStorage above are [ceph-0].

Closes-Bug: #1998649
Change-Id: I0405f6729b492af011531642a2ee374bb210fa82