Fix issues with connectivity check
The connectivity check has a sed command that currently
returns no values. Subsequently, this check currently does
not actually do anything. The error is outlined on the following
bz: https://bugzilla.redhat.com/show_bug.cgi?id=2150682

The patch resolves the sed issue, fixes handling of the MTU and methods
used to set_facts for each of these. Currently the split results in
empty values being returned, so we end up with the defaults of lo
and -28 for the MTU on all nodes.

Resolves: rhbz#2150682
Change-Id: I5cffea45c60bb98dbff2473a51e823f6b0ca174b