fix to be able to edit ns_records of /etc/designate/pools.yaml
As the RFC 1034 mentions, parent's NS records and child's NS records
must be the same.
For example:

  Parent:
    parent.example.com.       SOA ......
    child NS nameserver1.child.parent.example.com.
    child NS nameserver2.child.parent.example.com.
    child NS nameserver3.child.parent.example.com.

  Child:
    child.parent.example.com. SOA ......
          NS nameserver1.child.parent.example.com.
          NS nameserver2.child.parent.example.com.
          NS nameserver3.child.parent.example.com.

Currently, ns_records of /etc/designate/pools.yaml always start with
nsX and are not changeable. To make them the same as parent's NS
records, we have to be able to change ns_records to any value.

As ns_records should end with ".", this change makes sure that it ends
with ".".

Partial-Bug: 1999134
Change-Id: Id4466cfb01cf69dca06c56f4cd03146d8e38d363