Do not sudo stat Ceph spec file on Ansible host
When the "Stat spec file on ansible host" task is run
from an Ansible runner container, the task fails with:

"module_stderr": "/bin/sh: line 1: sudo: command not found\n",

This patch removes "become: true" so that the task
no longer fails under these conditions.

It shouldn't be necessary to use sudo to stat this
file on a standard undercloud or in an Ansible runner
container. There are other tasks which use become
and other tasks which delegate to localhost, but
this task in the tripleo_cephadm role which uses both.

Change-Id: I1aab213206d32ffae2f9fea92970a4bf13801f16
(cherry picked from commit 6c94f7e7f3fb84917021a4c2654ac6779607a84f)